move language-picker to shared components folder
